requirements.txt
alphabet_detector
beautifulsoup4
bleach
cachetools
CurrencyConverter
emoji
feedparser
future
geopy
googletrans==4.0.0-rc1
google-trans-new==1.1.9
humanize
jikanpy
lxml
nltk
Pillow>=4.2.0
pretty_errors
psutil
psycopg2-binary
pynewtonmath
pyrate-limiter
python-telegram-bot==13.6
regex
requests
spamwatch
speedtest-cli
spongemock
sqlalchemy
telethon==1.16.4
ujson
wikipedia
zalgo-text
fontTools
pyrogram
nekos.py
telegraph
tgCrypto>=1.2.2
youtube_dl
youtube_search
pytube
youtube_search_python
pymongo
motor
dnspython
redis
gtts
cloudscraper
wget
tswift
lyricsgenius
envparse
cryptg
cryptography
python-arq
aioredis
odmantic
aiohttp
pykeyboard
aiofiles
hachoir
carbonnow
bing_image_downloader
search_engine_parser
pytz
html2text
glitch_this
NumPy
opencv-python-headless
aiofiles
uvloop
python-dotenv
gpytranslate
glitch_this
NumPy
opencv-python-headless
countryinfo
faker
PyDictionary==2.0.1
ffmpeg
multicolorcaptcha
python-rapidjson
heroku3
git+https://github.com/python-telegram-bot/ptbcontrib.git
git+https://github.com/starry69/python-markdown2.git
coffeehouse
cowpy
selenium
httpx[http2]>=0.18.1,<1.0.0
hiredis
